Rearrange Kalibro Processor's tests
 * Now each class has its own test file
