+++ tests/processor/test_base.py
@@ -1,0 +1,15 @@
+from unittest import TestCase
+from nose.tools import assert_equal
+from mock import patch
+
+import kalibro_client
+from kalibro_client.processor.base import Base
+
+
+class TestProcessorBase(TestCase):
+    @patch('kalibro_client.config')
+    def test_service_address(self, kalibro_client_config):
+        kalibro_client_config.return_value = kalibro_client.DEFAULT_CONFIG
+
+        assert_equal(Base.service_address(), kalibro_client.DEFAULT_CONFIG['processor_address'])
+        kalibro_client_config.assert_called_once()
